+++ backend/etcd/id.go
@@ -9,7 +9,7 @@ import (
 	"sync/atomic"
 
 	"github.com/sensu/sensu-go/backend/store"
-	"go.etcd.io/etcd/client/v3"
+	clientv3 "go.etcd.io/etcd/client/v3"
 )
 
 var (
@@ -83,7 +83,9 @@ func (b *BackendIDGetter) keepAliveLease(ctx context.Context) {
 				}
 				return
 			}
+			LeaseOperationsCounter.WithLabelValues("sensu-etcd", LeaseOperationTypeKeepalive, LeaseOperationStatusAlive).Inc()
 			if resp.ID == clientv3.NoLease {
+				// I believe this to be impossible
 				b.errors <- errors.New("no lease")
 			}
 		case <-ctx.Done():
@@ -95,6 +97,7 @@ func (b *BackendIDGetter) keepAliveLease(ctx context.Context) {
 func (b *BackendIDGetter) getLease() (int64, <-chan *clientv3.LeaseKeepAliveResponse, error) {
 	// Grant a lease for 60 seconds
 	resp, err := b.client.Grant(b.ctx, backendIDLeasePeriod)
+	LeaseOperationsCounter.WithLabelValues("sensu-etcd", LeaseOperationTypeGrant, LeaseStatusFor(err)).Inc()
 	if err != nil {
 		return 0, nil, fmt.Errorf("error creating backend ID: error granting lease: %s", err)
 	}
@@ -105,12 +108,14 @@ func (b *BackendIDGetter) getLease() (int64, <-chan *clientv3.LeaseKeepAliveResp
 	value := fmt.Sprintf("%x", leaseID)
 	key := path.Join(backendIDKeyPrefix, value)
 	_, err = b.client.Put(b.ctx, key, value, clientv3.WithLease(leaseID))
+	LeaseOperationsCounter.WithLabelValues("sensu-etcd", LeaseOperationTypePut, LeaseStatusFor(err)).Inc()
 	if err != nil {
 		return 0, nil, fmt.Errorf("error creating backend ID: error creating key: %s", err)
 	}
 
 	// Keep the lease alive
 	ch, err := b.client.KeepAlive(b.ctx, leaseID)
+	LeaseOperationsCounter.WithLabelValues("sensu-etcd", LeaseOperationTypeKeepalive, LeaseStatusFor(err)).Inc()
 
 	return int64(leaseID), ch, err
 }