+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 - Added objectives (0.5, 0.9, 0.99) to `graphql_duration_seconds` metric.
 - Added `graphql_duration_seconds`, `graphql_duration_seconds_sum` &
 `graphql_duration_seconds_count` to the metrics log.
+- Added new prometheus metrics for tracking lease operations.
+- Added `sensu_go_lease_ops` to the metrics log.
 
 ### Fixed
 - Duration metrics for assets, pipeline, and eventd have been updated to use
@@ -65,7 +67,6 @@ metric.
 ## [6.5.1] - 2021-10-18
 
 ### Added
-- Added new prometheus metrics for tracking lease operations.
 - Added pipeline workflow handler processing counters
 
 ## [6.5.0] - 2021-10-12